pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "brain-mri-tumor-ensemble"
version = "0.1.0"
description = "Brain MRI tumor ensemble training and evaluation scripts"
authors = [{name = "Karlo Nahro"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "tensorflow>=2.12",
    "numpy>=1.26",
    "pyyaml>=6.0",
    "matplotlib>=3.7",
    "seaborn>=0.13",
    "scikit-learn>=1.3",
    "opencv-python-headless>=4.8",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["brain_mri_tumor_ensemble"]
